lab2 finished

+++ lab2/DebugPractice/DebugExercise1.java
@@ -2,10 +2,11 @@
  * Exercise for learning how the debug, breakpoint, and step-into
  * feature work.
  */
+
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+        double quotient = (double)top / (double)bottom;
+        int result = (int) Math.round(quotient);
         return result;
     }
